removed unwanted files, made a common file
