+++ packages/bruno-app/src/components/ResponsePane/TestResultsLabel/index.js
@@ -1,13 +1,28 @@
 import React from 'react';
 import { IconCircleCheck, IconCircleX } from '@tabler/icons';
 
-const TestResultsLabel = ({ results, assertionResults, preRequestTestResults, postResponseTestResults }) => {
+// Counts the tests of whichever phases the protocol has: HTTP/GraphQL in pre-request &
+// post-response plus its own tests and assertions, a gRPC call in its four call phases.
+const TestResultsLabel = ({
+  results,
+  assertionResults,
+  preRequestTestResults,
+  postResponseTestResults,
+  beforeCallStartTestResults,
+  beforeMessageSendTestResults,
+  afterMessageReceiveTestResults,
+  afterCallEndTestResults
+}) => {
   results = results || [];
   assertionResults = assertionResults || [];
   preRequestTestResults = preRequestTestResults || [];
   postResponseTestResults = postResponseTestResults || [];
+  beforeCallStartTestResults = beforeCallStartTestResults || [];
+  beforeMessageSendTestResults = beforeMessageSendTestResults || [];
+  afterMessageReceiveTestResults = afterMessageReceiveTestResults || [];
+  afterCallEndTestResults = afterCallEndTestResults || [];
 
-  if (!results.length && !assertionResults.length && !preRequestTestResults.length && !postResponseTestResults.length) {
+  if (!results.length && !assertionResults.length && !preRequestTestResults.length && !postResponseTestResults.length && !beforeCallStartTestResults.length && !beforeMessageSendTestResults.length && !afterMessageReceiveTestResults.length && !afterCallEndTestResults.length) {
     return 'Tests';
   }
 
@@ -23,8 +38,20 @@ const TestResultsLabel = ({ results, assertionResults, preRequestTestResults, po
   const numberOfPostResponseTests = postResponseTestResults.length;
   const numberOfFailedPostResponseTests = postResponseTestResults.filter((result) => result.status === 'fail').length;
 
-  const totalNumberOfTests = numberOfTests + numberOfAssertions + numberOfPreRequestTests + numberOfPostResponseTests;
-  const totalNumberOfFailedTests = numberOfFailedTests + numberOfFailedAssertions + numberOfFailedPreRequestTests + numberOfFailedPostResponseTests;
+  const numberOfBeforeCallStartTests = beforeCallStartTestResults.length;
+  const numberOfFailedBeforeCallStartTests = beforeCallStartTestResults.filter((result) => result.status === 'fail').length;
+
+  const numberOfBeforeMessageSendTests = beforeMessageSendTestResults.length;
+  const numberOfFailedBeforeMessageSendTests = beforeMessageSendTestResults.filter((result) => result.status === 'fail').length;
+
+  const numberOfAfterMessageReceiveTests = afterMessageReceiveTestResults.length;
+  const numberOfFailedAfterMessageReceiveTests = afterMessageReceiveTestResults.filter((result) => result.status === 'fail').length;
+
+  const numberOfAfterCallEndTests = afterCallEndTestResults.length;
+  const numberOfFailedAfterCallEndTests = afterCallEndTestResults.filter((result) => result.status === 'fail').length;
+
+  const totalNumberOfTests = numberOfTests + numberOfAssertions + numberOfPreRequestTests + numberOfPostResponseTests + numberOfBeforeCallStartTests + numberOfBeforeMessageSendTests + numberOfAfterMessageReceiveTests + numberOfAfterCallEndTests;
+  const totalNumberOfFailedTests = numberOfFailedTests + numberOfFailedAssertions + numberOfFailedPreRequestTests + numberOfFailedPostResponseTests + numberOfFailedBeforeCallStartTests + numberOfFailedBeforeMessageSendTests + numberOfFailedAfterMessageReceiveTests + numberOfFailedAfterCallEndTests;
 
   return (
     <div className="flex items-center">
